+++ packages/server/graphql/mutations/helpers/safeEndRetrospective.ts
@@ -165,10 +165,7 @@ const safeEndRetrospective = async ({
   )
   const timelineEventId = events[0]!.id
   const pg = getKysely()
-  await Promise.all([
-    pg.insertInto('TimelineEvent').values(events).execute(),
-    r.table('TimelineEvent').insert(events).run()
-  ])
+  await pg.insertInto('TimelineEvent').values(events).execute()
 
   if (team.isOnboardTeam) {
     const teamLeadUserId = await r